+++ src/applications/check-in/day-of/pages/Confirmation/MultipleAppointments.jsx
@@ -9,7 +9,7 @@ import BackToHome from '../../components/BackToHome';
 import BackToAppointments from '../../components/BackToAppointments';
 import Footer from '../../components/Footer';
 import AppointmentLocation from '../../components/AppointmentDisplay/AppointmentLocation';
-import TravelPayReimbursementLink from '../../components/TravelPayReimbursementLink';
+import TravelPayReimbursementLink from '../../../components/TravelPayReimbursementLink';
 
 const MultipleAppointments = props => {
   const { appointments, selectedAppointment, triggerRefresh } = props;
